requirements.txt
42cc-pystyle==0.0.16
atomicwrites==1.4.0
attrs==19.3.0
backports.functools-lru-cache==1.6.1
configparser==4.0.2
contextlib2==0.6.0.post1
Django==1.6.7
django-widget-tweaks==1.4.8
djangorestframework==3.0.0
flake8==2.6.2
freezegun==0.3.15
funcsigs==1.0.2
importlib-metadata==1.7.0
mccabe==0.5.3
more-itertools==5.0.0
packaging==20.4
pathlib2==2.3.5
Pillow==6.2.2
pluggy==0.13.1
py==1.9.0
pycodestyle==2.0.0
pyflakes==1.2.3
pyparsing==2.4.7
pytest==4.6.11
pytest-freezegun==0.4.1
python-dateutil==2.8.1
scandir==1.10.0
six==1.15.0
South==0.7.4
wcwidth==0.2.5
zipp==1.2.0

